feat: move merge logic to storage provider
