+++ lib/Onyx.js
@@ -296,7 +296,7 @@ function getCachedCollection(collectionKey) {
 function keysChanged(collectionKey, partialCollection) {
     // We are iterating over all subscribers similar to keyChanged(). However, we are looking for subscribers who are subscribing to either a collection key or
     // individual collection key member for the collection that is being updated. It is important to note that the collection parameter cane be a PARTIAL collection
-    // and does not represent all of the combined keys and values for a collection key. It is just the "new" data that was merged in via mergeCollection().
+    // and does not represent all of the combined keys and values for a collection key. It is just the "new" data that was qd in via mergeCollection().
     const stateMappingKeys = _.keys(callbackToStateMapping);
     for (let i = 0; i < stateMappingKeys.length; i++) {
         const subscriber = callbackToStateMapping[stateMappingKeys[i]];
@@ -905,58 +905,13 @@ function multiSet(data) {
         .catch(error => evictStorageAndRetry(error, multiSet, data));
 }
 
-// Key/value store of Onyx key and arrays of values to merge
-const mergeQueue = {};
-
 /**
  * @private
  * @param {String} key
  * @returns {Boolean}
  */
 function hasPendingMergeForKey(key) {
-    return Boolean(mergeQueue[key]);
-}
-
-/**
- * Given an Onyx key and value this method will combine all queued
- * value updates and return a single value. Merge attempts are
- * batched. They must occur after a single call to get() so we
- * can avoid race conditions.
- *
- * @private
- * @param {String} key
- * @param {*} data
- *
- * @returns {*}
- */
-function applyMerge(key, data) {
-    const mergeValues = mergeQueue[key];
-    if (_.isArray(data) || _.every(mergeValues, _.isArray)) {
-        // Array values will always just concatenate
-        // more items onto the end of the array
-        return _.reduce(mergeValues, (modifiedData, mergeValue) => [
-            ...modifiedData,
-            ...mergeValue,
-        ], data || []);
-    }
-
-    if (_.isObject(data) || _.every(mergeValues, _.isObject)) {
-        // Object values are merged one after the other
-        return _.reduce(mergeValues, (modifiedData, mergeValue) => {
-            // lodash adds a small overhead so we don't use it here
-            // eslint-disable-next-line prefer-object-spread, rulesdir/prefer-underscore-method
-            const newData = Object.assign({}, fastMerge(modifiedData, mergeValue));
-
-            // We will also delete any object keys that are undefined or null.
-            // Deleting keys is not supported by AsyncStorage so we do it this way.
-            // Remove all first level keys that are explicitly set to null.
-            return _.omit(newData, (value, finalObjectKey) => _.isNull(mergeValue[finalObjectKey]));
-        }, data || {});
-    }
-
-    // If we have anything else we can't merge it so we'll
-    // simply return the last value that was queued
-    return _.last(mergeValues);
+    return Storage.hasPendingMergeForKey(key);
 }
 
 /**
@@ -982,28 +937,7 @@ function applyMerge(key, data) {
  * @returns {Promise}
  */
 function merge(key, value) {
-    if (mergeQueue[key]) {
-        mergeQueue[key].push(value);
-        return Promise.resolve();
-    }
-
-    mergeQueue[key] = [value];
-    return get(key)
-        .then((data) => {
-            try {
-                const modifiedData = applyMerge(key, data);
-
-                // Clean up the write queue so we
-                // don't apply these changes again
-                delete mergeQueue[key];
-
-                return set(key, modifiedData);
-            } catch (error) {
-                Logger.logAlert(`An error occurred while applying merge for key: ${key}, Error: ${error}`);
-            }
-
-            return Promise.resolve();
-        });
+    return Storage.mergeItem(key, value);
 }
 
 /**
