+++ lib/storage/providers/AsyncStorage.js
@@ -57,7 +57,12 @@ const provider = {
      */
     multiMerge(pairs) {
         const stringPairs = _.map(pairs, ([key, value]) => [key, JSON.stringify(value)]);
+
         return AsyncStorage.multiMerge(stringPairs);
+    },
+
+    mergeItem(key, value) {
+        return this.multiMerge([[key, value]]);
     },
 
     /**
