+++ lib/storage/providers/LocalForage.js
@@ -8,10 +8,56 @@ import localforage from 'localforage';
 import _ from 'underscore';
 import SyncQueue from '../../SyncQueue';
 import fastMerge from '../../fastMerge';
+import * as Logger from '../../Logger';
 
 localforage.config({
     name: 'OnyxDB',
 });
+
+// Key/value store of Onyx key and arrays of values to merge
+const mergeQueue = {};
+
+/**
+ * Given an Onyx key and value this method will combine all queued
+ * value updates and return a single value. Merge attempts are
+ * batched. They must occur after a single call to get() so we
+ * can avoid race conditions.
+ *
+ * @private
+ * @param {String} key
+ * @param {*} data
+ *
+ * @returns {*}
+ */
+function applyMerge(key, data) {
+    const mergeValues = mergeQueue[key];
+    if (_.isArray(data) || _.every(mergeValues, _.isArray)) {
+        // Array values will always just concatenate
+        // more items onto the end of the array
+        return _.reduce(mergeValues, (modifiedData, mergeValue) => [
+            ...modifiedData,
+            ...mergeValue,
+        ], data || []);
+    }
+
+    if (_.isObject(data) || _.every(mergeValues, _.isObject)) {
+        // Object values are merged one after the other
+        return _.reduce(mergeValues, (modifiedData, mergeValue) => {
+            // lodash adds a small overhead so we don't use it here
+            // eslint-disable-next-line prefer-object-spread, rulesdir/prefer-underscore-method
+            const newData = Object.assign({}, fastMerge(modifiedData, mergeValue));
+
+            // We will also delete any object keys that are undefined or null.
+            // Deleting keys is not supported by AsyncStorage so we do it this way.
+            // Remove all first level keys that are explicitly set to null.
+            return _.omit(newData, (value, finalObjectKey) => _.isNull(mergeValue[finalObjectKey]));
+        }, data || {});
+    }
+
+    // If we have anything else we can't merge it so we'll
+    // simply return the last value that was queued
+    return _.last(mergeValues);
+}
 
 const provider = {
     /**
@@ -57,10 +103,36 @@ const provider = {
      * @return {Promise<void>}
      */
     multiMerge(pairs) {
-        const tasks = _.map(pairs, ([key, value]) => this.setItemQueue.push({key, value, shouldMerge: true}));
+        const tasks = _.map(pairs, ([key, value]) => {
+            if (mergeQueue[key]) {
+                mergeQueue[key].push(value);
+                return Promise.resolve();
+            }
+
+            mergeQueue[key] = [value];
+            Storage.getItem(key).then((data) => {
+                try {
+                    const modifiedData = applyMerge(key, data);
+
+                    // Clean up the write queue so we
+                    // don't apply these changes again
+                    delete mergeQueue[key];
+
+                    this.setItemQueue.push({key, modifiedData, shouldMerge: true}).then(() => Promise.resolve());
+                } catch (error) {
+                    Logger.logAlert(`An error occurred while applying merge for key: ${key}, Error: ${error}`);
+                }
+
+                return Promise.resolve();
+            });
+        });
 
         // We're returning Promise.resolve, otherwise the array of task results will be returned to the caller
         return Promise.all(tasks).then(() => Promise.resolve());
+    },
+
+    mergeItem(key, value) {
+        return this.multiMerge([[key, value]]);
     },
 
     /**
@@ -111,6 +183,10 @@ const provider = {
      */
     setItem(key, value) {
         return this.setItemQueue.push({key, value});
+    },
+
+    hasPendingMergeForQueue(key) {
+        return Boolean(mergeQueue[key]);
     },
 };
