+++ lib/storage/providers/SQLiteStorage.js
@@ -82,16 +82,20 @@ const provider = {
     multiMerge(pairs) {
         // Note: We use `ON CONFLICT DO UPDATE` here instead of `INSERT OR REPLACE INTO`
         // so the new JSON value is merged into the old one if there's an existing value
-        const query = `INSERT INTO keyvaluepairs (record_key, valueJSON) 
-             VALUES (:key, JSON(:value)) 
-             ON CONFLICT DO UPDATE 
-             SET valueJSON = JSON_PATCH(valueJSON, JSON(:value)); 
+        const query = `INSERT INTO keyvaluepairs (record_key, valueJSON)
+             VALUES (:key, JSON(:value))
+             ON CONFLICT DO UPDATE
+             SET valueJSON = JSON_PATCH(valueJSON, JSON(:value));
         `;
         const queryArguments = _.map(pairs, (pair) => {
             const value = JSON.stringify(pair[1]);
             return [pair[0], value];
         });
         return db.executeBatchAsync([[query, queryArguments]]);
+    },
+
+    mergeItem(key, value) {
+        return this.multiMerge([[key, value]]);
     },
 
     /**
